backened/requirements.txt

Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0


psycopg2-binary==2.9.11


firebase-admin==6.2.0
phonenumbers==8.13.27
python-jose[cryptography]==3.3.0
bcrypt==4.1.2


cloudinary==1.36.0
python-dotenv==1.0.0
Pillow==10.4.0

python-dateutil==2.8.2
django-cors-headers==4.2.0
django-environ==0.10.0


django-extensions==3.2.3